getHilfeSolution/modules/sendMQTTmessage/requirements.txt
azure-iot-device~=2.0.0
paho.mqtt


